refactor(folder structure): introducing actions folder
